backend/requirements.txt
fastapi==0.98.0
flake8==6.0.0
pydantic==1.10.7
pytest==7.2.2
pytest-asyncio==0.21.1
uvicorn==0.21.1
sqlalchemy==2.0.9
alembic==1.10.3
asyncpg==0.27.0
python-dotenv==0.19.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
celery[redis]==5.2.7
flower==1.2.0
redis==4.5.4
asyncio==3.4.3
gunicorn==20.1.0
bandit==1.7.5
httpx==0.23.1
black==23.7.0
httpx